+++ next-docs/components/sidebar/navigation.ts
@@ -48,6 +48,7 @@ const navigation = [
         ],
       },
       { name: 'RadioButton', href: '/core/radioButton' },
+      { name: 'Search', href: '/core/search' },
       { name: 'Select', href: '/core/select' },
       { name: 'Snackbar', href: '/core/snackbar' },
       { name: 'Switch', href: '/core/switch' },