+++ packages/select/CHANGELOG.md
@@ -1,5 +1,18 @@
 # @heathmont/moon-select
 
+## 9.21.0
+
+### Minor Changes
+
+- Add Search
+
+### Patch Changes
+
+- Updated dependencies [undefined]
+  - @heathmont/moon-core@9.21.0
+  - @heathmont/moon-themes@9.21.0
+  - @heathmont/moon-utils@9.21.0
+
 ## 9.20.1
 
 ### Patch Changes